Testing threads

+++ Current/Distributed Cloud-based Ray Tracing/Raytracer/camera.h
@@ -14,11 +14,11 @@ __device__ vec3 random_in_unit_disk(curandState* local_rand_state) {
 
 class camera {
 public:
-	__device__ camera(vec3 lookfrom, vec3 lookat, vec3 vup, double vfov, double aspect, double aperture, double focus_dist) { // vfov is top to bottom in degrees
+	__device__ camera(vec3 lookfrom, vec3 lookat, vec3 vup, float vfov, float aspect, float aperture, float focus_dist) { // vfov is top to bottom in degrees
 		lens_radius = aperture / 2.0f;
-		double theta = vfov * ((double)M_PI) / 180.0f;
-		double half_height = tan(theta / 2.0f);
-		double half_width = aspect * half_height;
+		float theta = vfov * ((float)M_PI) / 180.0f;
+		float half_height = tan(theta / 2.0f);
+		float half_width = aspect * half_height;
 		origin = lookfrom;
 		w = unit_vector(lookfrom - lookat);
 		u = unit_vector(cross(vup, w));
@@ -27,7 +27,7 @@ public:
 		horizontal = 2.0f * half_width * focus_dist * u;
 		vertical = 2.0f * half_height * focus_dist * v;
 	}
-	__device__ ray get_ray(double s, double t, curandState* local_rand_state) {
+	__device__ ray get_ray(float s, float t, curandState* local_rand_state) {
 		vec3 rd = lens_radius * random_in_unit_disk(local_rand_state);
 		vec3 offset = u * rd.x() + v * rd.y();
 		return ray(origin + offset, lower_left_corner + s * horizontal + t * vertical - origin - offset);
@@ -38,7 +38,7 @@ public:
 	vec3 horizontal;
 	vec3 vertical;
 	vec3 u, v, w;
-	double lens_radius;
+	float lens_radius;
 };
 
 #endif
